+++ app/src/main/java/com/example/gymprogresstracker/analysis/ProgressiveOverloadCalculator.java
@@ -11,7 +11,8 @@ public class ProgressiveOverloadCalculator {
     List<String> dateList = new ArrayList<>();
     HashMap<String, List<List<String>>> dataMap = new HashMap<>();
     List<List<String>> data = new ArrayList<>();
-    public ProgressiveOverloadCalculator(List<List<String>> data){
+
+    public ProgressiveOverloadCalculator(List<List<String>> data) {
         /*
         Daily Progressive Overload = Dimax - Di-1max / Di-1max)  *  100
 
@@ -23,33 +24,39 @@ public class ProgressiveOverloadCalculator {
 
         getResList();
     }
-    public List<Float> getResList(){
+
+    public List<Float> getResList() {
         List<Float> res = new ArrayList<>();
         res.add(0f);
-        for(int i = 1; i < dateList.size(); i++){
+        for (int i = 1; i < dateList.size(); i++) {
             float Wi = getWeight(dataMap.get(dateList.get(i)));
-            float Wi_1 = getWeight(dataMap.get(dateList.get(i-1)));
-            float progressiveOverLoad = (float) ((Wi - Wi_1)/(float) Wi_1)*100f;
+            float Wi_1 = getWeight(dataMap.get(dateList.get(i - 1)));
+            float progressiveOverLoad = ((Wi - Wi_1) / Wi_1) * 100f;
             res.add(progressiveOverLoad);
         }
         Log.e("", res.toString());
         return res;
     }
-    public float getWeight(List<List<String>> data){
+
+    public float getWeight(List<List<String>> data) {
         float weight = 0f;
         float reps = 0;
-        for(List<String> record: data){
+        for (List<String> record : data) {
             weight += Float.parseFloat(record.get(2));
             reps += Float.parseFloat(record.get(3));
         }
-        return weight/reps;
+        return weight / reps;
     }
-    public void parseDataByDate(){
+
+    public void parseDataByDate() {
         int I = 0;
-        for(List<String> record:data){
+        for (List<String> record : data) {
             String date = record.get(0);
-            if(dateList.size() == 0){dateList.add(date);dataMap.put(date, new ArrayList<>());}
-            if(!dateList.get(dateList.size()-1).equals(date)){
+            if (dateList.size() == 0) {
+                dateList.add(date);
+                dataMap.put(date, new ArrayList<>());
+            }
+            if (!dateList.get(dateList.size() - 1).equals(date)) {
                 dateList.add(date);
                 dataMap.put(date, new ArrayList<>());
             }
@@ -57,7 +64,8 @@ public class ProgressiveOverloadCalculator {
         }
         Log.e("e", dataMap.toString());
     }
-    public List<Float> getResult(){
+
+    public List<Float> getResult() {
         return getResList();
     }
 }
